Move CommandEvaluation into command.py
