+++ src/statue/command.py
@@ -4,13 +4,50 @@ import importlib
 import os
 import subprocess  # nosec
 import sys
-from dataclasses import dataclass, field
-from typing import List, Optional
+from dataclasses import asdict, dataclass, field
+from typing import Any, Dict, List, Optional
 
 import pkg_resources
 
 from statue.exceptions import CommandExecutionError
 from statue.verbosity import DEFAULT_VERBOSITY, is_silent, is_verbose
+
+
+@dataclass
+class CommandEvaluation:
+    """Evaluation result of a command."""
+
+    command: "Command"
+    success: bool
+
+    def as_json(self) -> Dict[str, Any]:
+        """
+        Return command evaluation as json dictionary.
+
+        :return: Self as dictionary
+        :rtype: Dict[str, Any]
+        """
+        command_json = {
+            key: value
+            for key, value in asdict(self.command).items()
+            if value is not None
+        }
+        return dict(command=command_json, success=self.success)
+
+    @classmethod
+    def from_json(cls, command_evaluation: Dict[str, Any]) -> "CommandEvaluation":
+        """
+        Read command evaluation from json dictionary.
+
+        :param command_evaluation: Json command evaluation
+        :type command_evaluation: Dict[str, Any]
+        :return: Parsed command evaluation
+        :rtype: CommandEvaluation
+        """
+        return CommandEvaluation(
+            command=Command(**command_evaluation["command"]),
+            success=command_evaluation["success"],
+        )
 
 
 @dataclass
